requirements.txt
argon2-cffi==20.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.5
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
fugashi==1.1.0
gensim==3.7.0
idna==2.10
ipadic==1.0.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mecab-python3==1.0.4
mistune==0.8.4
mosestokenizer==1.1.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.1
notebook==6.4.3
numpy==1.20.2
openfile==0.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plac==1.3.3
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
regex==2021.3.17
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.8.0
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
smart-open==5.0.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.1.0
tokenizers==0.10.1
toolwrapper==2.1.0
torch==1.8.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.4.2
typing-extensions==3.7.4.3
uctools==1.3.0
unidic==1.0.3
unidic-lite==1.0.8
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
